requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
pandas
openpyxl